backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
chromadb==0.4.18
pypdf2==3.0.1
pydantic==2.5.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.2
tiktoken==0.5.2
